Fix regex to accept -1 as "valid" sequence number
The corresponding tests to this code do not need to be updated because
they mock out the behaviour of retrieving the sequence number. The tests
themselves were passing and expecting the actual implementation to
return a -1 sequence number. However, due to this bug (that this commit
fixes), the tests were passing as a false positive. This is now
corrected.