Connect to mysql over unix socket
This component is a local "sidecar" that connects to MySQL and reports
status and provides other features around Galera observability and
bootstrapping.

All connections are local and this component already had access to the
MySQL unix socket in pxc-release.  This simplifies the process of
enabling "require-secure-transport" for the pxc-mysql job.

[#177349859](https://www.pivotaltracker.com/story/show/177349859)